Possibly the dumbest thing I've ever written
The only encoder port that reached the elevator encoder after I removed the octoquad is the ExpansionHub motor encoder port 0. Port 0 is plugged into the RightDriveMotor. As a result, that *elevator* encoders position will be reported as rightDriveMotor.getPosition(). This position is obvisouly needed within the elevator class, so I passed the position function into the elevator class via the constructor and an IntSupplier. This would not be a problem if i could access that encoder like actually. Thanks FTC!

- Removed octoquad.
- Added elevator encoder back.
- Lost many braincells.
- TODO: Remove intake slide.